requirements.txt
# ARIA — Python Dependencies
# Install: pip install -r requirements.txt

# ── Core ──────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ── Database ──────────────────────────────────────────────────────────
sqlalchemy>=2.0.0
alembic>=1.13.0           # DB migrations
psycopg2-binary>=2.9.0    # PostgreSQL adapter

# ── ML & Data ─────────────────────────────────────────────────────────
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
xgboost>=2.0.0
prophet>=1.1.5
pyarrow>=14.0.0           # Parquet support

# ── LLM & Agent ───────────────────────────────────────────────────────
anthropic>=0.34.0
langchain>=0.3.0
langchain-anthropic>=0.2.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0

# ── Data Ingestion ────────────────────────────────────────────────────
requests>=2.31.0
serpapi>=0.1.5            # SerpAPI: competitor prices + Google Trends (replaces pytrends)

# ── Cache ─────────────────────────────────────────────────────────────
redis>=5.0.0              # Optional — falls back to file cache

# ── Monitoring & Observability ────────────────────────────────────────
langfuse>=2.0.0           # LLM tracing (optional)
prometheus-client>=0.19.0 # Metrics

# ── API & Dashboard ──────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
streamlit>=1.32.0
httpx>=0.27.0              # async HTTP client for FastAPI

# ── Scheduling ────────────────────────────────────────────────────────
schedule>=1.2.0           # Simple job scheduler

# ── Testing ───────────────────────────────────────────────────────────
pytest>=7.4.0
pytest-asyncio>=0.23.0

# ── Utilities ─────────────────────────────────────────────────────────
python-dateutil>=2.8.2
tabulate>=0.9.0           # Pretty table output
colorlog>=6.8.0           # Colored logging